fix(sema)!: a value composite owns heap iff a member does
What was: `ownsHeap` (both sema legs) and HIR's `normCtx.ownsHeap` answered
YES for every value composite -- the v1 rule, "a Copy composite owns its
box" -- long after the storage flips made a struct, tuple, union or fixed
array inline. `rejectArmHandingOutBorrowedPayload` (SEM3197) gates on that
axis, so `Hold(p) => p` over `&Held` was refused for
`@copy type Pair = { a: int, b: int }`, a payload nobody could double-free
(RV2-DEBT-256); every struct of ints, every `Option<int>` and
`TaskResult<int>` carried a drop obligation, and MIR emitted a drop for each
that reclaimed nothing on either lane.

What is: one structural walk, `ownsHeapIn` (ownership_axes.go), answers for
`tc.ownsHeap`, `Result.OwnsHeap` and -- exported as `sema.OwnsHeapIn` -- for
HIR's leg: a reference-counted scalar owns its block, a borrow owns nothing,
a value composite owns heap iff a field, tuple element, union tag payload or
fixed-array element does (recursively, through `own`, fail-closed on a shape
the walk cannot read), and a handle owns heap iff it is not Copy, as before.
The backend's `typeOwnsHeap` already walked; the three compiler legs now say
what it says.

Fixtures that modelled "a droppable" as a struct of ints were carried by the
old rule and are adapted, not weakened: `Held` in the ternary test owns a
string; the MIR ownership fixtures' `Cell` is `{ a: int, b: float }`, so the
clone, alias and duplicated-scrutinee rows still have something to release;
the MIR harness marks its fixture `Task` as a runtime handle by name, as the
driver marks core's -- without the mark `Task<T> = { __opaque: int }` is a
struct of one int to the walk, where the composite rule used to hide the
omission. `TestCapabilityDroppableIsNotOwnsHeap` pinned a disagreement the
storage flip was expected to end; it is
`TestCapabilityDroppableAgreesWithOwnsHeap` now and pins the agreement from
both sides.

Goldens (Rule 12): seventeen files under mir/, mono/ and hir_borrow/ change
by one pattern -- a local whose type is a composite of scalars loses
`[owns_heap]`, its `drop`, and the `tmp_retval` copy that only sequenced that
drop before the return; `owned_temp(...)` around an awaited
`TaskResult<nothing>` goes for the same reason. Every removed drop was a
no-op: inline storage on the VM (scratch rewinds at the instruction
boundary), an empty glue body natively. Two new sema fixtures:
`valid/arm_hands_out_copy_scalar_payload.sg` (empty .diag) and
`invalid/ownership/arm_hands_out_payload_owning_a_string.sg` (one SEM3197).

Falsifiers, red with the composite branch of the walk reverted to
`return true`:

  TestOwnsHeapFollowsTheMembers
      Pair: OwnsHeap=true, want false   (and Plain, Outer, HeldPair,
      ArrayFixed<int, const 3, 3>, (int, int) -- both legs)
  TestArmHandingOutPayloadFollowsTheOwnsHeapAxis/copy_pair_of_ints_is_handed_out
      unexpected sema diagnostics: [SEM3197] cannot hand `p` out of this arm
  TestCapabilityDroppableAgreesWithOwnsHeap
      drop obligations treat a Copy struct of two integers as owning heap
  TestOwnershipAxesAgreeWithCopyToday
      type 25 (struct): OwnsHeap=true, want false (IsCopy=true)

and the old binary on the new valid fixture:

  error SEM3197 testdata/golden/sema/valid/arm_hands_out_copy_scalar_payload.sg:14:20
  cannot hand `p` out of this arm: the compare only borrows the value it was
  matched against

The float probe is the axis's own answer: a `@copy` pair of floats owns two
counted blocks, so SEM3197 keeps refusing it (a row in the same test).
Whether the rule's escape for a payload that took its own reference should
widen from bare scalars to a Copy composite whose extraction retains is a
rule question, left open.

+++ internal/mir/tag_payload_moveout_test.go
@@ -22,7 +22,12 @@ tag Payload(string);
 tag Empty();
 type Slot = Payload(string) | Empty;
 
-@copy type Cell = { a: int, b: int };
+// Cell owns heap through its float: a composite owns what its members own, so a
+// pair of ints would be bits, the cloned envelope would have nothing to
+// release, and the borrowed and duplicated shapes below would rightly collapse
+// into one. The distinction this test pins is only observable on an envelope
+// with something in it.
+@copy type Cell = { a: int, b: float };
 tag Held(Cell);
 tag Absent();
 @copy type Holder = Held(Cell) | Absent;
